changes -  diseño responsive  para registro - solucion de errores, registro multiples al validar , add diseño de ticket  para ver puntos

+++ src/pages/private/admin/Setting/Portafolio/Servicios/Servicios.jsx
@@ -89,6 +89,8 @@ const Servicios = () => {
   };
 
   const handleDeleteService = (id) => {
+    let confirmationEnabled = true;
+
     modals.openConfirmModal({
       title: "Eliminacion de Servicio",
       centered: true,
@@ -99,9 +101,12 @@ const Servicios = () => {
       confirmProps: { color: "red" },
 
       onConfirm: () => {
-        dispatch(deleteServicio(id));
-        Notify("Eliminacion Exitosa", "", "success");
-        handleCloseAction();
+        if (confirmationEnabled) {
+          confirmationEnabled = false;
+          dispatch(deleteServicio(id));
+          Notify("Eliminacion Exitosa", "", "success");
+          handleCloseAction();
+        }
       },
     });
   };
